+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to
 - ⬆️(backend) upgrade celery to version 5.6.3
 - ⚡️(backend) stop using LEFT(value, LENGTH(path)) in sql queries
 - 🚚(project) switch docspec image to ghcr.io/docspec/api #2553
+- 🚚(global) move favorite documents API endpoint 
+  to `/documents/favorites/` #2540
 
 ### Fixed